requirements.txt
requests
pre-commit
python-dotenv
aiofiles
pillow
pytesseract
distlib==0.3.4
fastapi==0.75.0
gunicorn==20.1.0
identify==2.4.12
Jinja2==3.1.1
pre-commit==2.17.0
pydantic
python-multipart
pytest==7.1.1
PyYAML==6.0
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.14.0